+++ sdks/java/io/pulsar/src/main/java/org/apache/beam/sdk/io/pulsar/PulsarSourceDescriptor.java
@@ -5,23 +5,32 @@ import com.google.auto.value.AutoValue;
 import org.apache.beam.sdk.schemas.AutoValueSchema;
 import org.apache.beam.sdk.schemas.annotations.DefaultSchema;
 import org.apache.beam.sdk.schemas.annotations.SchemaFieldName;
+import org.apache.beam.sdk.values.PInput;
 import org.apache.pulsar.client.api.MessageId;
 
 import java.io.Serializable;
 
 @DefaultSchema(AutoValueSchema.class)
 @AutoValue
-public abstract class PulsarSource implements Serializable {
+public abstract class PulsarSourceDescriptor implements Serializable {
+
+    @SchemaFieldName("topic")
+    abstract String getTopic();
 
     @SchemaFieldName("start_offset")
     @Nullable
     abstract Long getStartOffset();
 
-    @SchemaFieldName("topic")
-    abstract String getTopic();
 
-    @SchemaFieldName("startReadMessageId")
-    abstract MessageId getStartMessageId();
 
+    public static PulsarSourceDescriptor of(
+            String topic,
+            Long startOffsetTimestamp
+    ) {
+        return new AutoValue_PulsarSourceDescriptor(
+                topic,
+                startOffsetTimestamp
+        );
+    }
 
 }
